# backend/openapi.yaml
openapi: 3.0.0
info:
  title: SafeWalk API
  version: 1.0.0
  description: API REST para el sistema SafeWalk U de rutas seguras para estudiantes.
servers:
  - url: http://localhost:3000/api
    description: Servidor local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RutaRequest:
      type: object
      required:
        - nombre_ruta
        - nivel_seguridad
        - tiempo_estimado
      properties:
        nombre_ruta:
          type: string
          example: Ruta UTPL
        descripcion:
          type: string
          example: Ruta segura hacia la universidad
        nivel_seguridad:
          type: string
          enum: [BAJO, MEDIO, ALTO]
          example: ALTO
        tiempo_estimado:
          type: number
          example: 15
...
paths:
...
  /routes:
    get:
      tags:
        - Rutas
      summary: Obtener todas las rutas seguras
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de rutas

    post:
      tags:
        - Rutas
      summary: Crear ruta segura
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RutaRequest"
      responses:
        "201":
          description: Ruta creada correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod
  /routes/{id}:
    get:
      tags:
        - Rutas
      summary: Obtener ruta por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ruta encontrada
        "404":
          description: Ruta no encontrada

    put:
      tags:
        - Rutas
      summary: Actualizar ruta segura
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RutaRequest"
      responses:
        "200":
          description: Ruta actualizada correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

    delete:
      tags:
        - Rutas
      summary: Eliminar ruta segura
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ruta eliminada correctamente
        "403":
          description: No autorizado
...